rename: 일관성을 위한 폴더명 변경
