+++ src/pages/snack-list/index.tsx
