+++ src/routes/routes.tsx
@@ -1,15 +1,15 @@
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import Home from "../pages/home/page";
+import Home from "../pages/home";
 import App from "../app";
-import SnackList from "../pages/snack-list/page";
-import Snack from "../pages/snack-list/snack/page";
-import NotFound from "../pages/not-found/page";
-import FoodList from "../pages/food-list/page";
-import Food from "../pages/food-list/food/page";
+import SnackList from "../pages/snack-list";
+import Snack from "../pages/snack-list/snack";
+import NotFound from "../pages/not-found";
+import FoodList from "../pages/food-list";
+import Food from "../pages/food-list/food";
 import Login from "../pages/login";
-import Signup from "../pages/signup/page";
-import Mypage from "../pages/mypage/page";
-import MealRegistration from "../pages/meal-registration/page";
+import Signup from "../pages/signup";
+import Mypage from "../pages/mypage";
+import MealRegistration from "../pages/meal-registration";
 
 const router = createBrowserRouter([
   {
